[pam_post_auth] Split code from header
